frames: Add new X11 frames client
This separate X11 client takes care of creating frames for client
windows, Mutter will use this client to delegate window frame
rendering and event handling.

This client is itself a GTK4 application, using real GtkWindows to
implement window frames, the client window is expected to be
reparented by the WM (still Mutter) into the GtkWindow X11 Window,
and picked up from there.

In order to make things behave, the frames need to forward some
messaging to client windows (e.g. WM_DELETE_WINDOW when the X button
is pressed), and some things it needs to read and synchronize with
the client (e.g. window titles).

Other than that, this is still a CSD GTK4 window, so window drags
and resizes, and window context menus are forwarded for the WM to
handle.